+++ src/index.js
@@ -1,0 +1,3 @@
+import createProject from './utils/createProject';
+
+export { createProject };